--- lolstaticdata/lolstaticdata/items/__init__.py
@@ -1,4 +1,0 @@
-# from lolstaticdata.items import __main__
-#
-#
-# __main__.rewrite()
