--- lolstaticdata/pyproject.toml
@@ -1,5 +1,0 @@
-[tool.black]
-line-length = 120
-target-version = ['py37']
-include = '\.pyi?$'
-exclude = ''
