--- lolstaticdata/requirements.txt
@@ -1,8 +1,0 @@
-beautifulsoup4
-dataclasses-json
-lxml
-natsort
-requests
-slpp
-fuzzywuzzy
-stringcase
