+++ src/main/resources/templates/index.html
@@ -802,9 +802,15 @@
     </div>
     <!-- about us 모달창 E -->
 
-    <script th:src="@{/js/test.js}"></script> <!-- 모든 기능 이벤트 -->
+<!--    <script th:src="@{/js/test.js}"></script> &lt;!&ndash; 모든 기능 이벤트 &ndash;&gt;-->
     <script th:src="@{/js/decimal.js}"></script> <!-- 계산식 -->
     <script th:src="@{/js/web.js}"></script> <!-- 웹페이지 이벤트 관련 -->
+    <script th:src="@{/js/statusL.js}"></script>
+    <script th:src="@{/js/skillL.js}"></script>
+    <script th:src="@{/js/itemL.js}"></script>
+    <script th:src="@{/js/statusR.js}"></script>
+    <script th:src="@{/js/skillR.js}"></script>
+    <script th:src="@{/js/itemR.js}"></script>
     <!-- 메인영역 E -->
 
     <!-- 푸터영역 -->
